requirements.txt
streamlit
pandas
openai
numpy
scikit-learn
python-dotenv